# app/Models/Inventario.php
<?php

namespace App\Models;

use Illuminate\Database\Eloquent\Factories\HasFactory;
use Illuminate\Database\Eloquent\Model;

class Inventario extends Model
{
    use HasFactory;
    protected $table = 'inventario';
    protected $fillable = [
        'producto',
        'descripcion',
        'cantidad',
        'unidad',
        'nivel_reorden',
        'costo_por_unidad',
        'ubicacion_almacen',
        'proveedor_id',
    ];

    public function proveedor()
    {
        return $this->belongsTo(Proveedor::class);
    }
}