# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Cafetería",
        "description": "API para gestión de cafetería",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor Local"
        }
    ],
    "paths": {
...
        "/api/inventarios": {
            "get": {
                "tags": [
                    "Inventarios"
                ],
                "summary": "Obtener lista de inventarios",
                "operationId": "721655bcda437b9623a493da77e2fd43",
                "responses": {
                    "200": {
                        "description": "Lista completa de inventarios",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Inventario"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Inventarios"
                ],
                "summary": "Crear un nuevo inventario",
                "operationId": "e2544a669fb3cce9342d9cf42dfbf205",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "producto",
                                    "cantidad",
                                    "unidad",
                                    "proveedor_id"
                                ],
                                "properties": {
                                    "producto": {
                                        "type": "string",
                                        "example": "Café en grano"
                                    },
                                    "cantidad": {
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "unidad": {
                                        "type": "string",
                                        "example": "kg"
                                    },
                                    "proveedor_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Inventario creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Inventario"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Datos inválidos"
                    }
                }
            }
        },
        "/api/inventarios/{id}": {
            "get": {
                "tags": [
                    "Inventarios"
                ],
                "summary": "Obtener un inventario por ID",
                "operationId": "61ca5382ad56e1c47831198872094a4a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del inventario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del inventario",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Inventario"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Inventario no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Inventarios"
                ],
                "summary": "Actualizar un inventario existente",
                "operationId": "dd96c5e900591f5e268aa18a8320e524",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del inventario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "producto": {
                                        "type": "string",
                                        "example": "Café en grano"
                                    },
                                    "cantidad": {
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "unidad": {
                                        "type": "string",
                                        "example": "kg"
                                    },
                                    "proveedor_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Inventario actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Inventario"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Inventario no encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Inventarios"
                ],
                "summary": "Eliminar un inventario",
                "operationId": "72192e1c345bc6be025c87d0b8b860d4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del inventario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Inventario eliminado exitosamente"
                    },
                    "404": {
                        "description": "Inventario no encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Inventario": {
                "required": [
                    "producto",
                    "cantidad",
                    "unidad",
                    "proveedor_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "producto": {
                        "type": "string",
                        "example": "Café en grano"
                    },
                    "cantidad": {
                        "type": "integer",
                        "example": 100
                    },
                    "unidad": {
                        "type": "string",
                        "example": "kg"
                    },
                    "proveedor_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}